Wrote up slope experiment
Adjusted experiment in inclinacion_de_la_base to set up a new flowline model with steeper slope.  Students should be able to modify the new slope as they wish and explore output.

Next need to replace single plot with side-by-side plots, so that profiles on different bedrock will be legible.