Added test of width-dependent climate response
Two experiments in reassion_variable: (1) students compare the equilibrium length and volume for different ELAs on three glaciers of different width; (2) students apply a step change in climate to the three glaciers and compare how much their volume changes in a given time.

(1) may be incorrectly coded - was confusing when I was doing it.  Revisit.