Fixed width experiment
Added line so that glacier width is adjusted only in the accumulation area for anchura_del_glaciar experiment.  FM points out that if the width were adjusted over the whole glacier (as first written), the equilibrium lengths would be identical and it would not be intuitive for students.  Glaciers with the same tongue width but different accumulation areas should be a better illustration.