Added experiments to NB 6
Set up cell where students can manipulate the accumulation area width and slope mix of three glaciers imagined to be in the same hydrologic 'basin' (but not intersecting).  Added experiments similar to NB 5, reaccion_variable, in which students examine equilibrium volume and response to a step change in ELA, but this time for an aggregate collection of different glaciers.